Add ui key to state
