+++ src/reducers/index.js
@@ -1,70 +1,9 @@
 import { combineReducers } from 'redux';
 
-import InitialState from 'constants/InitialState';
-
-import loading from 'reducers/loading';
+import ui from 'reducers/ui';
 import oba from 'reducers/oba';
 
-import {
-  SET_USER_LOCATION,
-  SET_USER_LOCATION_ERROR,
-  SET_CURRENT_ROUTE,
-  SET_CURRENT_TAB,
-  SET_GLOBAL_ERROR,
-} from 'constants/ActionTypes';
-
-
-function currentTab(state = InitialState.currentTab, action = {}) {
-  if (action.type === SET_CURRENT_TAB) {
-    return action.payload;
-  }
-  return state;
-}
-
-function currentAgency(state = InitialState.currentAgency) {
-  return state;
-}
-
-function currentRoute(state = InitialState.currentRoute, action = {}) {
-  if (action.type === SET_CURRENT_ROUTE) {
-    return action.payload;
-  }
-  return state;
-}
-
-function userLocation(state = null, action = {}) {
-  if (action.type === SET_USER_LOCATION) {
-    return action.payload;
-  }
-  return state;
-}
-
-function userLocationError(state = null, action = {}) {
-  if (action.type === SET_USER_LOCATION_ERROR) {
-    return action.payload;
-  }
-  return state;
-}
-
-function recent(state = InitialState.recent, action = {}) {
-  return state;
-}
-
-function globalError(state = InitialState.globalError, action = {}) {
-  if (action.type === SET_GLOBAL_ERROR) {
-    return action.payload;
-  }
-  return state;
-}
-
 export default combineReducers({
-  currentTab,
-  currentAgency,
-  currentRoute,
-  userLocation,
-  userLocationError,
-  globalError,
-  recent,
-  loading,
+  ui,
   oba,
 });
