+++ src/reducers/ui/loading.js
@@ -8,21 +8,21 @@ import {
   SET_STOPS_FOR_ROUTE_LOADING,
 } from 'constants/ActionTypes';
 
-function allRoutesLoading(state = InitialState.loading.allRoutesLoading, action = {}) {
+function allRoutesLoading(state = InitialState.ui.loading.allRoutesLoading, action = {}) {
   if (action.type === SET_ALL_ROUTES_LOADING) {
     return action.payload;
   }
   return state;
 }
 
-function nearbyTripsLoading(state = InitialState.loading.nearbyTripsLoading, action = {}) {
+function nearbyTripsLoading(state = InitialState.ui.loading.nearbyTripsLoading, action = {}) {
   if (action.type === SET_NEARBY_TRIPS_LOADING) {
     return action.payload;
   }
   return state;
 }
 
-function stopsForRouteLoading(state = InitialState.loading.stopsForRouteLoading, action = {}) {
+function stopsForRouteLoading(state = InitialState.ui.loading.stopsForRouteLoading, action = {}) {
   if (action.type === SET_STOPS_FOR_ROUTE_LOADING) {
     return action.payload;
   }
